move metaeq_cli into metaseq
